Finished refactoring the macro to make it more modular
