+++ flint-macros/src/codegen/util.rs
@@ -1,0 +1,5 @@
+pub fn generate_unique_id() -> u32 {
+    use std::sync::atomic::{AtomicU32, Ordering};
+    static COUNTER: AtomicU32 = AtomicU32::new(0);
+    COUNTER.fetch_add(1, Ordering::Relaxed)
+}
