+++ flint-macros/src/widget/util.rs
@@ -1,0 +1,5 @@
+use syn::Ident;
+
+pub fn is_layout_widget(name: &Ident) -> bool {
+    name.to_string() == "Layout"
+}
